+++ .github/workflows/run-tests.yml
@@ -25,6 +25,9 @@ jobs:
             - name: Checkout the branch
               uses: actions/checkout@v3
 
+            - name: Install Poetry
+              uses: snok/install-poetry@v1
+            
             - name: Use Node.js ${{ matrix.node-version }}
               uses: actions/setup-node@v3
               with:
