--- things/calculator/http/flask/test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-main=$(poetry run python ./main.py) & ../../../../node_modules/mocha/bin/mocha.js --exit "*.test.js";
-pid="$(lsof -i :5000 | grep -i python | awk '{print $2}')"
-kill -9 $pid;
