+++ things/calculator/http/flask/test/mocha.opts
@@ -1,0 +1,2 @@
+--exit
+--timeout 5000
