+++ things/calculator/mqtt/js/test/mocha.opts
@@ -1,0 +1,2 @@
+--exit
+--timeout 5000
